Move errors in one file
