+++ index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-const { Terminus } = require('./lib/terminus')
-const { HealthCheckError } = require('./lib/errors/health-check-error')
-const { TerminusError } = require('./lib/errors/terminus-error')
-const { HealthCheckNotFoundError } = require('./lib/errors/health-check-not-found-error')
+const { Terminus } = require('./lib/terminus');
+const { TerminusError, HealthCheckError, HealthCheckNotFoundError } = require('./lib/terminus-error')
 
 module.exports = {
   Terminus,
