--- lib/errors/health-check-not-found-error.js
@@ -1,3 +1,0 @@
-const { TerminusError } = require('./terminus-error')
-
-module.exports.HealthCheckNotFoundError = class HealthCheckNotFoundError extends TerminusError {}
