--- lib/errors/terminus-error.js
@@ -1,1 +1,0 @@
-module.exports.TerminusError = class TerminusError extends Error { }
