+++ lib/terminus-error.spec.js
@@ -1,6 +1,6 @@
 const expect = require('chai').expect
 
-const { HealthCheckError } = require('./errors/health-check-error')
+const { HealthCheckError } = require('./terminus-error')
 
 describe('Terminus error', () => {
   it('saves the cause', () => {
