+++ lib/terminus.js
@@ -4,7 +4,7 @@ const stoppable = require('stoppable')
 const {
   promisify
 } = require('es6-promisify')
-const { HealthCheckNotFoundError } = require('./errors/health-check-not-found-error')
+const { HealthCheckNotFoundError } = require('./terminus-error');
 
 function noopResolves () {
   return Promise.resolve()
