+++ lib/terminus.spec.js
@@ -6,7 +6,7 @@ const expect = require('chai').expect
 const fetch = require('node-fetch')
 
 const { Terminus } = require('./terminus')
-const { HealthCheckError } = require('./errors/health-check-error')
+const { HealthCheckError } = require('./terminus-error')
 
 describe('Terminus', () => {
   let server
@@ -281,43 +281,41 @@ describe('Terminus', () => {
         healthChecks: {
           '/health': () => Promise.resolve()
         }
-      });
+      })
 
-      expect(await terminus.getHealthStatus('/health')).to.deep.eq({ status: 'ok' });
+      expect(await terminus.getHealthStatus('/health')).to.deep.eq({ status: 'ok' })
     })
     it('returns the correct health status with info if everything is ok', async () => {
-      const healthStatus = { ram: 8200 };
+      const healthStatus = { ram: 8200 }
       const terminus = new Terminus(server, {
         healthChecks: {
           '/health': () => Promise.resolve(healthStatus)
         }
-      });
+      })
 
-      expect(await terminus.getHealthStatus('/health')).to.deep.eq({ status: 'ok', info: healthStatus });
+      expect(await terminus.getHealthStatus('/health')).to.deep.eq({ status: 'ok', info: healthStatus })
     })
     it('returns the correct health status if a error is thrown', async () => {
-      const errorMessage = 'Error message';
+      const errorMessage = 'Error message'
       const terminus = new Terminus(server, {
         healthChecks: {
           '/health': () => Promise.reject(new Error(errorMessage))
         }
-      });
+      })
 
-      expect(await terminus.getHealthStatus('/health')).to.deep.eq({ status: 'error', error: [errorMessage] });
+      expect(await terminus.getHealthStatus('/health')).to.deep.eq({ status: 'error', error: [errorMessage] })
     })
 
     it('returns the correct causes if a HealthCheckError is thrown', async () => {
-      const errorMessage = 'Error message';
-      const causes = ['Error1', 'Error2'];
+      const errorMessage = 'Error message'
+      const causes = ['Error1', 'Error2']
       const terminus = new Terminus(server, {
         healthChecks: {
           '/health': () => Promise.reject(new HealthCheckError(errorMessage, causes))
         }
-      });
+      })
 
-      expect(await terminus.getHealthStatus('/health')).to.deep.eq({ status: 'error', error: causes });
+      expect(await terminus.getHealthStatus('/health')).to.deep.eq({ status: 'error', error: causes })
     })
   })
-
-
 })
